+++ src/components/main/Main.styled.ts
@@ -1,4 +1,5 @@
 import { theme } from '@components/utils/styles/theme';
+import { StyleType } from '@interfaces/style';
 import styled from 'styled-components';
 
 export const ContentContainer = styled.div`
@@ -20,4 +21,17 @@ export const FeedContainer = styled.div`
   justify-content: space-between;
   align-items: center;
   background-color: ${theme.color.main};
+`;
+
+export const Content = styled.div`
+  width: 100%;
+  flex: 5;
+  display: flex;
+  flex-direction: ${(props: StyleType) => props.flexDirection || 'row'};
+`;
+
+export const ContentItem = styled.div`
+  width: 100%;
+  height: 100%;
+  border: 1px solid ${theme.color.main};
 `;
